compiles but definitely not done

+++ lastProject/error.h
@@ -26,7 +26,7 @@ unsigned int currConfig;
 
 float error_ir(turret_comm_t*);
 float error_sonar(turret_comm_t*);
-float error_tx(create_comm_t*, float, int);
+float error_tx(create_comm_t*, float, float);
 float error_ta(create_comm_t*, float);
 int* WhatDoISee(turret_comm_t*);
